Show a warning message when creating Tracker objects on MacOS
As MacOS support is currently experimental until we gather enough
confidence that everything works and that we can support the platform in
the long term, show a warning message when Tracker objects are created.

Additionally, raise an exception if "native_traces" is used as a Tracker
option as we don't support that feature in MacOS yet.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>