Stop allocating a new vector for each unwind call
Up until now, each time we would unwind the native stack we would start
with an empty vector, and then reserve enough space for it to hold the
largest stack we had seen so far on the current thread. Instead of
starting from scratch each time, save this vector in a thread-local
variable so that we can reuse it from one unwind call to the next.

Use `pthread_setspecific` instead of C++'s `thread_local` specifier, as
we may need to create this vector while TLS variables are in the process
of being destroyed (if a TLS variable's destructor makes an allocation,
it calls our malloc hook, which will want to create this vector). When
a C++ TLS variable is created after TLS variables for a thread have
already been destroyed, libstdc++ will cause that variable's destructor
to fire on already freed memory in some arbitrary thread in the future.
When the same condition occurs with a `pthread_setspecific` TLS
variable, POSIX guarantees that another round of TLS destruction will
occur. After some maximum number of rounds (but at least 4), libpthread
will give up, but in that case our failure mode is just leaking a small
amount of memory (a vector whose size is proportional to the depth of
the native call stack of a destructor call that made an allocation).

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>
Signed-off-by: Pablo Galindo <pablogsal@gmail.com>